Solve sensor update problem
The root cause was the drdyVal variable in transfer_data_to_drdy_vars
function. The variable was local. Seems like the value located at
the address of drdyVal variable was being update. This is possible
since the variable should not be in use after the function
transfer_data_to_drdy_vars exits. However, dma controller was still
using the address. Solution was to define drdyVal as a static variable.